+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Changed
+- Extend getting dao delegates by the specific delegation type
+
 ## [0.3.7] - 2026-02-02
 
 ### Added